TASK-3: Created a directory for vehicle types, put program logic into a separate class
