+++ TASK-3/VehicleTypes/Minivan.cs
