+++ TASK-3/VehicleTypes/Truck.cs
